Fix for new @import statements being ignored
Adding an @import statement in a .less file fails to update the list of imports for that file, which means that any subsequent change to the imported file(s) does not cause the .less file with @import statements to be recompiled. There is an open issue for this in the original gulp-watch-less.